move compressor, pump, consumer system to infrastructure
