+++ src/libecalc/infrastructure/energy_components/compressor/__init__.py
