+++ src/libecalc/infrastructure/energy_components/compressor/component.py
