+++ src/libecalc/infrastructure/energy_components/consumer_system.py
