+++ src/libecalc/infrastructure/energy_components/factory.py
@@ -2,11 +2,11 @@ from typing import Union, overload
 
 from libecalc.common.component_type import ComponentType
 from libecalc.common.time_utils import Period
-from libecalc.core.consumers.compressor import Compressor
-from libecalc.core.consumers.pump import Pump
 from libecalc.core.models.compressor import create_compressor_model
 from libecalc.core.models.pump import create_pump_model
 from libecalc.infrastructure.components import CompressorComponent, PumpComponent
+from libecalc.infrastructure.energy_components.compressor import Compressor
+from libecalc.infrastructure.energy_components.pump import Pump
 
 
 @overload
